Move SolicitorCreate handlers to their own package
