+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/solicitorcreatecase/SolicitorCreateConsentedMidHandler.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.handler;
+package uk.gov.hmcts.reform.finrem.caseorchestration.handler.solicitorcreatecase;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -6,31 +6,36 @@ import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.controllers.GenericAboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.finrem.caseorchestration.handler.CallbackHandler;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
-import uk.gov.hmcts.reform.finrem.caseorchestration.service.OnStartDefaultValueService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.ConsentOrderService;
 
+import java.util.List;
 import java.util.Map;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class SolicitorCreateConsentedAboutToStartHandler implements CallbackHandler<Map<String, Object>> {
+public class SolicitorCreateConsentedMidHandler implements CallbackHandler<Map<String, Object>> {
 
-    private final OnStartDefaultValueService service;
+    private final ConsentOrderService consentOrderService;
 
     @Override
     public boolean canHandle(CallbackType callbackType, CaseType caseType, EventType eventType) {
-        return CallbackType.ABOUT_TO_START.equals(callbackType)
+        return CallbackType.MID_EVENT.equals(callbackType)
             && CaseType.CONSENTED.equals(caseType)
-            && (EventType.SOLICITOR_CREATE.equals(eventType));
+            && EventType.SOLICITOR_CREATE.equals(eventType);
     }
 
     @Override
     public GenericAboutToStartOrSubmitCallbackResponse<Map<String, Object>> handle(CallbackRequest callbackRequest,
                                                                                    String userAuthorisation) {
-        service.defaultCivilPartnershipField(callbackRequest);
+
+        log.info("Invoking Solicitor Create mid event");
+        List<String> errors = consentOrderService.performCheck(callbackRequest, userAuthorisation);
+
         return GenericAboutToStartOrSubmitCallbackResponse.<Map<String, Object>>builder()
-            .data(callbackRequest.getCaseDetails().getData()).build();
+            .data(callbackRequest.getCaseDetails().getData()).errors(errors).build();
     }
 }
